Replace ContainerSpec.Circularity with ContainerTests.Circularity.
